+++ apps/ledger-live-mobile/src/newArch/features/Market/hooks/useMarketCoinData.ts
@@ -24,7 +24,6 @@ export const useMarketCoinData = ({ currencyId }: HookProps) => {
   const { data: currency, isLoading } = useCurrencyData({
     counterCurrency,
     id: currencyId,
-    range,
   });
 
   return {
